backend/pyproject.toml
[project]
name = "neo-visualizer-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "yfinance>=0.2.40",
    "pandas>=2.2",
    "numpy>=1.26",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["domain", "application", "infrastructure", "presentation"]
